Remove slow git prompt colouring
Running 'git status' to determine the colour of the bash prompt
can take some time for large repos, which makes working in the
terminal irritating. The colour was also often out-of-date with
multiple terminals open in the same repo, so removing the git
status colouring will stop this misleading behaviour.